+++ src/main/java/com/example/EcommerceSpringBoot/gateways/IProductGateway.java
@@ -5,5 +5,5 @@ import com.example.EcommerceSpringBoot.dto.ProductDTO;
 import java.io.IOException;
 
 public interface IProductGateway {
-    ProductDTO getProductByIdGateway(int id) throws IOException;
+    ProductDTO getProductByIdGateway(Long id) throws IOException;
 }
